app/requirements.txt
streamlit==1.28.0
langchain-core==0.1.0
langchain-groq==0.1.0
langgraph==0.0.40
python-dotenv==1.0.0
requests==2.31.0
pypdf2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6